Template changes!
New massagetherapiespages controller to handle the links on the left
navigation for therapies. Changed routing to route anything with
therapies/blah to the new controller. If it has nothing after it, it is
handled by the massagepages controller.

Template changed to include /header.php before /controller/header.php if the
latter exists, and /footer.php after /controller/footer.php if the latter
exists. So you can add on to the headers and footers without copying code.

Still might need to modify the massagepagescontroller for therapies()
because I can just hard-code the variables in therapies.php instead of
passing them in via php.